MINIFICPP-1825 Create Properties at compile time
Also Relationships, Output Attributes and Dynamic Properties.

This only applies to Processor and Controller Service metadata used in
generating the manifest and PROCESSORS.md.  I did not change Properties and
Relationships used later on during runtime, as those are not always const
and sometimes they need to own their own strings.

There are a few changes in the generated PROCESSORS.md, mainly because the
list of allowed values for a Property is no longer alphabetized; now the
allowed values are listed in the order they appear in either the property
definition or in the corresponding smart enum.